# backend/storage/api-docs-admin/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Next Starter API — Admin",
        "description": "Documentação da área **Admin** (equipe interna / back-office) do projeto Laravel Next Starter.\n\nEsta documentação é restrita: o acesso à interface Swagger UI e ao JSON da spec exige autenticação HTTP Basic própria da documentação (usuário/senha definidos em `.env`), **independente** do login da API.\n\nApós acessar a documentação, os endpoints em si continuam exigindo o Bearer JWT obtido em `POST /api/admin/login`.",
        "contact": {
            "name": "Equipe de Desenvolvimento"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API (relativo ao host onde o Swagger UI está sendo servido)"
        }
    ],
    "paths": {
...
        "/admin/empresas/{empresaId}/enderecos": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Listar endereços da empresa",
                "operationId": "9a0f8a52df75999e5c7c04dd9586a2bc",
                "parameters": [
                    {
                        "name": "empresaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de endereços.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EmpresaEndereco"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Cadastrar endereço da empresa",
                "operationId": "05e76ef84bd26e2808c263f4b10336be",
                "parameters": [
                    {
                        "name": "empresaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo",
                                    "municipio_id",
                                    "principal",
                                    "ativo",
                                    "cep",
                                    "logradouro",
                                    "numero",
                                    "bairro"
                                ],
                                "properties": {
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "COMERCIAL",
                                            "FISCAL",
                                            "CORRESPONDENCIA",
                                            "COBRANCA",
                                            "ENTREGA"
                                        ]
                                    },
                                    "municipio_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "principal": {
                                        "description": "Só pode existir um endereço principal por empresa.",
                                        "type": "boolean"
                                    },
                                    "ativo": {
                                        "type": "boolean"
                                    },
                                    "cep": {
                                        "description": "Somente dígitos (a máscara é removida automaticamente).",
                                        "type": "string",
                                        "pattern": "^[0-9]{8}$",
                                        "example": "01310100"
                                    },
                                    "logradouro": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "numero": {
                                        "type": "string",
                                        "maxLength": 5
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço cadastrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EmpresaEndereco"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/empresas/{empresaId}/enderecos/{enderecoId}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Visualizar endereço da empresa",
                "operationId": "d46f152f22e9347cef7801a2dfc5095a",
                "parameters": [
                    {
                        "name": "empresaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "enderecoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EmpresaEndereco"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Atualizar endereço da empresa",
                "operationId": "2e6115ca1fb530872f818805893daca8",
                "parameters": [
                    {
                        "name": "empresaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "enderecoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo",
                                    "municipio_id",
                                    "principal",
                                    "ativo",
                                    "cep",
                                    "logradouro",
                                    "numero",
                                    "bairro"
                                ],
                                "properties": {
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "COMERCIAL",
                                            "FISCAL",
                                            "CORRESPONDENCIA",
                                            "COBRANCA",
                                            "ENTREGA"
                                        ]
                                    },
                                    "municipio_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "principal": {
                                        "description": "Só pode existir um endereço principal por empresa.",
                                        "type": "boolean"
                                    },
                                    "ativo": {
                                        "type": "boolean"
                                    },
                                    "cep": {
                                        "description": "Somente dígitos (a máscara é removida automaticamente).",
                                        "type": "string",
                                        "pattern": "^[0-9]{8}$",
                                        "example": "01310100"
                                    },
                                    "logradouro": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "numero": {
                                        "type": "string",
                                        "maxLength": 5
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EmpresaEndereco"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Excluir endereço da empresa",
                "operationId": "33005c619675326f747f859b07010a12",
                "parameters": [
                    {
                        "name": "empresaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "enderecoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/NoContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/empresas/{empresaId}/enderecos/{enderecoId}/ativar": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Ativar/reativar endereço da empresa",
                "operationId": "3ddf9cf5c19c32029f0e08c5aec67273",
                "parameters": [
                    {
                        "name": "empresaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "enderecoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço ativado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EmpresaEndereco"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmpresaEndereco": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "empresa_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "COMERCIAL",
                            "FISCAL",
                            "CORRESPONDENCIA",
                            "COBRANCA",
                            "ENTREGA"
                        ]
                    },
                    "tipo_descricao": {
                        "type": "string",
                        "example": "Comercial"
                    },
                    "municipio_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "ativo": {
                        "type": "boolean"
                    },
                    "principal": {
                        "type": "boolean"
                    },
                    "cep": {
                        "type": "string",
                        "example": "01310100"
                    },
                    "logradouro": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "numero": {
                        "type": "string",
                        "maxLength": 5
                    },
                    "bairro": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "complemento": {
                        "type": "string",
                        "maxLength": 50,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "municipio": {
                        "$ref": "#/components/schemas/Municipio"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Unauthorized": {
                "description": "Não autenticado: token ausente, inválido, expirado, ou sessão associada ao token não está mais ativa.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Autenticado, porém sem permissão para executar esta ação.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Registro não encontrado.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
...
            "NoContent": {
                "description": "Operação realizada com sucesso, sem conteúdo de retorno."
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via **JSON Web Token (JWT)**.\n\n1. Faça login em `POST /api/login` (Private) ou `POST /api/admin/login` (Admin).\n2. Envie o token retornado no header `Authorization: Bearer {token}` nas demais requisições.\nO token carrega um `session_id`, validado a cada requisição contra a sessão ativa do usuário (tabela de sessões).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}